Updates resume links and image path (#3)
* Updates resume links and image path

Updates the links to point to the "download/latest" release path.

Changes the resume image path to the latest version.

* Ensures `make update` always runs

Adds the `-B` flag to the `make update` command in the CI workflow.

This ensures that `make update` is always executed, even if no changes are detected, which guarantees dependencies are up-to-date.

* Makes 'update' target depend on 'build'

Ensures that the 'build' target is executed before the 'assets/current' target is updated.
This guarantees that the assets are up-to-date before the update process begins.

* Adds amd64-unknown-linux target

Adds the `amd64-unknown-linux` target to the rust toolchain.
This ensures that the project can be compiled for linux on amd64 architecture.

* Updates target triple for cross-compilation

Corrects the target triple specification in the rust-toolchain.toml
file to align with the expected format for cross-compilation.

* Actually read the docs

* Fiddling

* That was fun but lets be realistic here

* This is karma

* Font stuff

* Regen I guess

* Font paths

* Quotes

* Ruby stuff

* Regen

* Ignore the pdf asset

A "Modified" timestamp is stored in the PDF contents so compilation always produces a diff and breaks the clean workspace check

* Adds font management with Fontist

Introduces Fontist for managing fonts, ensuring consistent rendering across different environments.

Configures the build process to install required fonts before compiling the resume.

Adds ignore rules for Fontist's working directory.

* Simplifies CI workflow

Removes explicit font installation and arguments from the CI workflow.

This change simplifies the CI process by relying on default font configurations, leading to a cleaner and more maintainable workflow.

* Adds development documentation

Provides a TL;DR section for common development tasks.

Adds a pre-requisites section to guide developers on setting up
their environment.

* Regen with the italics and thin fonts

* Release workflow tweaks

* Revert "Ignore the pdf asset"

This reverts commit bcd919cb056a8b3c91764d2d15698b9c5be8698e.

* Simplify